% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-task.R
\name{knit_print.cm_latex_block}
\alias{knit_print.cm_latex_block}
\title{Knitr printing method for causalmetrics LaTeX blocks}
\usage{
\method{knit_print}{cm_latex_block}(x, ...)
}
\arguments{
\item{x}{A \code{cm_latex_block} object.}

\item{...}{Passed to \code{knitr::asis_output()}.}
}
\description{
Knitr printing method for causalmetrics LaTeX blocks
}
\keyword{internal}
