% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att-dose.R
\name{plot_att_dose}
\alias{plot_att_dose}
\title{Plot dose-response estimates}
\usage{
plot_att_dose(x, what = c("att", "acr"))
}
\arguments{
\item{x}{An object from \code{\link[=att_dose]{att_dose()}}.}

\item{what}{\code{"att"} (level effects) or \code{"acr"} (causal responses).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot dose-response estimates
}
