% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate-blp.R
\name{plot_cate_blp}
\alias{plot_cate_blp}
\title{Plot a best linear predictor of the CATE along one covariate}
\usage{
plot_cate_blp(x, newdata, x_var = names(newdata)[1], ...)
}
\arguments{
\item{x}{A \code{cm_blp} object.}

\item{newdata}{Data frame on which to evaluate the curve; the column
\code{x_var} is placed on the horizontal axis.}

\item{x_var}{Name of the covariate on the horizontal axis (default: the
first column of \code{newdata}).}

\item{...}{Passed to the \code{predict()} method of \code{cm_blp} objects (for instance \code{seed}).}
}
\value{
A ggplot object.
}
\description{
Plot a best linear predictor of the CATE along one covariate
}
