% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate-blp.R
\name{plot_cate_gate}
\alias{plot_cate_gate}
\title{Plot group average treatment effects}
\usage{
plot_cate_gate(x)
}
\arguments{
\item{x}{A \code{cm_gate} object.}
}
\value{
A ggplot object with pointwise (thick) and simultaneous (thin)
intervals.
}
\description{
Plot group average treatment effects
}
