% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate-reg.R
\name{plot_mediate_sensitivity}
\alias{plot_mediate_sensitivity}
\title{Plot a mediation sensitivity curve}
\usage{
plot_mediate_sensitivity(x)
}
\arguments{
\item{x}{A \code{cm_med_sensitivity} object.}
}
\value{
A ggplot object of the indirect effect against the error correlation.
}
\description{
Plot a mediation sensitivity curve
}
