% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mte.R
\name{plot_mte}
\alias{plot_mte}
\title{Plot a marginal treatment effect curve and its policy weights}
\usage{
plot_mte(x, what = c("curve", "weights"))
}
\arguments{
\item{x}{A \code{cm_mte} object.}

\item{what}{\code{"curve"} or \code{"weights"}.}
}
\value{
A ggplot object.
}
\description{
Plot a marginal treatment effect curve and its policy weights
}
