% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{plot_policy_tree}
\alias{plot_policy_tree}
\title{Draw a depth-limited policy tree}
\usage{
plot_policy_tree(x)
}
\arguments{
\item{x}{A \code{cm_policy} object with \code{method = "tree"} and the exhaustive
engine.}
}
\value{
A ggplot object.
}
\description{
Draw a depth-limited policy tree
}
