% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-aipw.R
\name{print.cm_aipw}
\alias{print.cm_aipw}
\title{Print method for \code{cm_aipw} objects}
\usage{
\method{print}{cm_aipw}(x, ...)
}
\arguments{
\item{x}{A \code{cm_aipw} object returned by \code{\link[=est_aipw]{est_aipw()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{cm_aipw} objects
}
\keyword{internal}
