% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-dml.R
\name{print.cm_dml}
\alias{print.cm_dml}
\title{Print method for \code{cm_dml} objects}
\usage{
\method{print}{cm_dml}(x, ...)
}
\arguments{
\item{x}{A \code{cm_dml} object returned by \code{\link[=est_dml]{est_dml()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{cm_dml} objects
}
\keyword{internal}
