% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.cm_detect_conflicts}
\alias{.cm_detect_conflicts}
\title{Detect function name conflicts between attached packages}
\usage{
.cm_detect_conflicts(pkgs)
}
\description{
Detect function name conflicts between attached packages
}
\keyword{internal}
